Remove position reset logic. Check win/loss/version first
The position reset logic is no longer necessary. It was
only necessary due to an error where I would scroll down
when I shouldn't before.

Check win/lose/version before player1/player2 characters.
If we scroll down and are waiting on the server to give
us the next results we could potentially grab the previous
p1/p2 and match those with the next win/lose result